Move the visualisations to their own classes
