+++ index.html
@@ -23,12 +23,16 @@
 </head>
 
 <body>
-
-  <p>Order: <select id="order">
+  <p>Order:
+    <select id="order">
       <option value="name">by Name</option>
       <option value="count">by Frequency</option>
       <option value="group">by Cluster</option>
-    </select></p>
+    </select>
+  </p>
+
+  <div id="adj-matrix"></div>
+  <div id="node-links"></div>
 </body>
 
 </html>
