+++ src/visualizations/visualization.ts
@@ -1,0 +1,7 @@
+import { Correspondants, Email } from "../data";
+
+export abstract class Visualization {
+
+    abstract visualize(emails: Email[], correspondants: Correspondants): Promise<void>
+
+}
